evil file for bots
If bots were to request any paths that do not actually exist they will encounter this neat little 941 byte file.
Although kind of inefficient it is hand-crafted and contains about 70 MB of payload after decompressing it with gzip twice, which is what the headers indicate that the client ought to do.
This file is returned in all cases even when the client has indicated it cannot handle gzip, but hey, if something's crawling my website is that *my* problem?
The payload are opening brackets which should throw any JSON parser off.

Signed-off-by: benaryorg <binary@benary.org>